Unimed_Scraping/requirements.txt
fastapi==0.111.0
uvicorn==0.34.0
selenium==4.18.1
supabase==2.12.0
python-dotenv==1.0.1
loguru==0.7.2
pydantic==2.6.3
requests==2.31.0
webdriver-manager==4.0.2
pandas==2.2.3